Fix compiler warning: implicit conversion loses integer precision
src/print_tree_view_list.c:399:63: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
                str_tmp = gsb_string_uniform_new_line ( text, strlen ( text ) );
                          ~~~~~~~~~~~~~~~~~~~~~~~~~~~         ^~~~~~~~~~~~~~~
src/print_tree_view_list.c:652:63: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
                str_tmp = gsb_string_uniform_new_line ( text, strlen ( text ) );
                          ~~~~~~~~~~~~~~~~~~~~~~~~~~~         ^~~~~~~~~~~~~~~

+++ src/utils_str.c
@@ -880,7 +880,7 @@ gchar *gsb_string_extract_int ( const gchar *chaine )
  *
  * \return chaine au format unix
  */
-gchar *gsb_string_uniform_new_line ( const gchar *chaine, gint nbre_char )
+gchar *gsb_string_uniform_new_line ( const gchar *chaine, gssize nbre_char )
 {
     gchar **tab_str = NULL;
     gchar *result = NULL;
