+++ src/utils_str.h
@@ -26,7 +26,7 @@ gchar *		gsb_string_remplace_string 								(const gchar *str,
 																	 gchar *new_str);
 gchar * 	gsb_string_truncate 									(gchar * string);
 gchar *		gsb_string_uniform_new_line 							(const gchar *chaine,
-																	 gint nbre_char);
+																	 gssize nbre_char);
 gchar * 	latin2utf8 												(const gchar * inchar);
 gchar *		limit_string 											(gchar *string,
 												                     gint length);
